Fix date sorting & UI & reload
`allItems` should always be in a sorted state to avoid constant , expensive resorting
`sectionDates` should be sorted by default for better readability
reload activities if scrolled to the end. Comments are always loaded in full. So, we refetch all comments in full.
Use a DistpatchGroup to avoid multiple sorts reloads with different, (partially) missing data.

Signed-off-by: Harshal Sengar <harshal.sengar@t-systems.com>